fix: remove `internal-types.ts`
